[fix]Qiu: fix express server async running issue
because of express server in grunt:test start in background so that
the karma:unit test will be skiped and will cause a phantomjs running
in background.

and the grunt:compile job will generate a big app.js file which is 1.7M,
during the file generating, grunt job will continue so that cannot find
angular in unit testing. stange issue !!!!